+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## 1.7.0.22 [Work-In-Progress]
-- Simplified source code (expressions, slider names, etc. are now loaded directly in the PARM resource).
+- Simplified source code:
+  * expressions, slider names, etc. are now loaded directly in the PARM resource).
+	* Load-methods don't return boolean and a reason string. Instead they return 0 (success) or a error message ID otherwise.
 - As side effect: The slider positions will be preserved when a filter is made
 
 ## 1.7.0.21 [23-Nov-2023]